Extract cobra command, svc service as a separate package
