--- cmd/builder/test/replaces.otel.yaml
@@ -1,36 +1,0 @@
-extensions:
-  zpages:
-
-receivers:
-  otlp:
-    protocols:
-      grpc:
-        endpoint: localhost:4317
-
-processors:
-  resourcedetection:
-    detectors:
-      - env
-  routing:
-    default_exporters:
-    - logging
-    from_attribute: X-Tenant
-    table:
-    - value: acme
-      exporters: 
-      - logging
-
-exporters:
-  logging:
-
-service:
-  extensions: [zpages]
-  pipelines:
-    traces:
-      receivers:
-        - otlp
-      processors:
-        - resourcedetection
-        - routing
-      exporters:
-        - logging
