+++ cmd/otelcol/collector_windows_test.go
@@ -15,7 +15,7 @@
 //go:build windows
 // +build windows
 
-package service
+package otelcol
 
 import (
 	"os"
@@ -30,13 +30,13 @@ import (
 	"go.opentelemetry.io/collector/internal/testcomponents"
 )
 
-func TestWindowsService_Execute(t *testing.T) {
+func TestNewSvcHandler(t *testing.T) {
 	os.Args = []string{"otelcol", "--config", filepath.Join("testdata", "otelcol-config.yaml")}
 
 	factories, err := testcomponents.DefaultFactories()
 	require.NoError(t, err)
 
-	s := NewWindowsService(CollectorSettings{BuildInfo: component.NewDefaultBuildInfo(), Factories: factories})
+	s := NewSvcHandler(CollectorSettings{BuildInfo: component.NewDefaultBuildInfo(), Factories: factories})
 
 	colDone := make(chan struct{})
 	requests := make(chan svc.ChangeRequest)
